inserisci commento e parte db per photo

+++ src/java/DoveCiboPK/Review.java
@@ -14,11 +14,11 @@ import java.sql.Date;
 public class Review {
 
     private Integer id;
-    private Double global_value;
-    private Double food;
-    private Double service;
-    private Double value_of_money;
-    private Double atmosphere;
+    private Integer global_value;
+    private Integer food;
+    private Integer service;
+    private Integer value_of_money;
+    private Integer atmosphere;
     private String name;
     private String description;
     private Date date_creation;
@@ -27,7 +27,7 @@ public class Review {
     private Integer validation;
     private Replies repile;
 
-    public Review(Integer id, Double global_value, Double food, Double service, Double value_of_money, Double atmosphere, String name, String description, Date date_creation, User creator, Photo photo, Integer validation) {
+    public Review(Integer id, Integer global_value, Integer food, Integer service, Integer value_of_money, Integer atmosphere, String name, String description, Date date_creation, User creator, Photo photo, Integer validation) {
         this.id = id;
         this.global_value = global_value;
         this.food = food;
@@ -47,7 +47,7 @@ public class Review {
         this.repile = repile;
     }
 
-    public Double getAtmosphere() {
+    public Integer getAtmosphere() {
         return atmosphere;
     }
 
@@ -63,15 +63,15 @@ public class Review {
         return name;
     }
 
-    public Double getFood() {
+    public Integer getFood() {
         return food;
     }
 
-    public Double getService() {
+    public Integer getService() {
         return service;
     }
 
-    public Double getValue_of_money() {
+    public Integer getValue_of_money() {
         return value_of_money;
     }
 
@@ -79,7 +79,7 @@ public class Review {
         return photo;
     }
 
-    public Double getGlobal_value() {
+    public Integer getGlobal_value() {
         return global_value;
     }
